+++ tests/Http/ResponseTest.php
@@ -2,6 +2,7 @@
 
 namespace Dingo\Api\Tests\Http;
 
+use stdClass;
 use Dingo\Api\Http\Response;
 use PHPUnit_Framework_TestCase;
 
@@ -14,5 +15,17 @@ class ResponseTest extends PHPUnit_Framework_TestCase
     public function testGettingInvalidFormatterThrowsException()
     {
         Response::getFormatter('json');
+    }
+
+
+    public function testNonCastableObjectsSetAsOriginalContent()
+    {
+        $object = new stdClass;
+        $object->id = 'test';
+
+        $response = new Response($object);
+
+        $this->assertNull($response->getContent());
+        $this->assertSame($object, $response->getOriginalContent());
     }
 }